+++ pydirectus/__init__.py
@@ -1,3 +1,17 @@
 from .directus import Directus
+from .collection import Collection
+from .folder import Folder
+from .field import Field
+from .file import File
+from .query import Query
+from .session import Session
 
-__all__ = ["Directus"]
+# export main object
+__all__ = ["Directus",
+           "Collection",
+           "Folder",
+           "Field",
+           "File",
+           "Query",
+           "Session"
+           ]